renamed groovyserver -> groovyserver.sh
'groovyclient.sh' is newly added and it must have a 'sh' extension to be distinguished from a native client.
If a script had a 'sh' extension and another one didn't in same directory, it would be ugly.
That's the reason. Anyway it's a issue of source code management.
When executables/dist, it will be renamed again and removed the 'sh' extension.