removed a commit id from suffix of jar file's name only for SNAPSHOT
because if classifier is assigned in doFirst, even though there is no different in both inputs and outputs.
due to the suffix, tasks after jar cannot skip as UP-TO-DATE even though set appropriate inputs/outputs configurations.
now the suffix isn't important very much, so I decided to remove it simply.